WIP: Load configuration, save map filters.
Filters selected can now be saved to Community Lands.
When enabled, agents can name and save filters, sharing
them with the community selected in the backend.

A config model is now loaded prior to rendering MapFilter,
which contains information regarding the location of the
geojson data, whether or not filters can be saved, and
default filters to load. Currently, only the settings for
the geojson file location and saving filters are respected.

Still to complete on this front:

* Load filters on start from config
* Support description when saving maps?
* Display summary of filter selections on save panel.