Fix translation issue with loading dates.
Dates are stored in their own language in plain text for
filters, which works fine when loading a filter saved in
one language in that same language. This can fail when
loading a filter in one language that's stored in a
different language.

Updated the loading process to parse the date information
first according to all available locales to choose the best
one to read it, then translate it to the app locale.