Additional details shown when saving maps.
When saving maps to community lands, and entity is returned
that yields additional information that can be displayed to
users, such as whether or not the operation was a new save
or an update, and the timestamp of the last data update for
that community.

Use this information to display when the last update was; if
no data exists, convey that to the user as well.